# bank-service/swagger.yaml
openapi: 3.0.0
info:
  title: Simple Bank API - Bank Management
  version: 1.0.0
  description: API for managing banks and bank accounts in the Simple Bank application
servers:
  - url: http://localhost:{port}/api/bank
    description: Development server
    variables:
      port:
        default: "3002"
paths:
...
  /accounts/{userId}:
    get:
      summary: Retrieve bank account(s) by user ID
      tags:
        - Bank Accounts
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
          description: The ID of the user whose accounts are to be retrieved
      responses:
        '200':
          description: Bank accounts fetched successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Bank accounts fetched successfully
                  accounts:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: number
                          example: 1
                        userId:
                          type: string
                          example: "123e4567-e89b-12d3-a456-426614174000"
                        bankId:
                          type: number
                          example: 1
                        balance:
                          type: number
                          format: float
                          example: 1000.50
        '400':
          description: Invalid user ID
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Invalid user ID
        '404':
          description: No accounts found for user
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: No accounts found for user
        '500':
          description: Error fetching bank accounts
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Error fetching bank accounts
                  error:
                    type: string
                    example: Unknown error
...